hopefully the last fix. bye guys. its been nice knowing ya.

+++ app/patient/deletePatient.php
@@ -1,22 +1,35 @@
+
 <?php
-     include_once('../database/conn.php');
+/* This code deletes a patient from the database. The deleteid variable is passed in by the patient.html page when the user clicks on the delete button. The code attempts to delete the patient from the database, and returns a success message if the delete is successful, and an error message if the delete fails. */
+include_once('../database/conn.php');
+error_reporting(E_ALL & ~E_WARNING & ~E_NOTICE);
 
-
-if(isset($_POST['deleteid'])){
+if (isset($_POST['deleteid'])) {
     //delete service
     $id = $_POST['deleteid'];
-    $sql = mysqli_query($conn,"DELETE FROM patients WHERE patient_id='$id'");
+    $chk = mysqli_query($conn, "SELECT * FROM appointments  WHERE patient_id='$id'");
 
-    if($sql){
-        $data = ['message'=>'success', 'status'=>200];
-                echo json_encode($data);
-                return ;
-        
-    }
-    else{
-        $data = ['message'=>'failed to delete patient', 'status'=>404];
+    if (mysqli_num_rows($chk) > 0) {
+        $data = ['message' => 'Error deleting patient: Constraint (Patient Exists in another table.)', 'status' => 404];
         echo json_encode($data);
-        return ;
-    }
+        return;
+    } else {
+        $sql = mysqli_query($conn, "DELETE FROM patients WHERE patient_id='$id'");
+
+        if ($sql) {
+            $data = ['message' => 'success', 'status' => 200];
+            echo json_encode($data);
+            return;
+        } else {
+            $data = ['message' => 'Failed To delete Pateint.', 'status' => 404];
+            echo json_encode($data);
+            return;
+        }
+    } 
+} else {
+    $msg =  "Error deleting patient: " . "No post data received";
+    $data = ['message' => "$msg", 'status' => 404];
+    echo json_encode($data);
+    return;
 }
 ?>
